backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
apscheduler
playwright
PyJWT
httpx
python-dotenv
